* Move system code maps to parsers.system * Convert simple system parsers to use code maps directly
