+++ aiopioneer/parsers/system.py
@@ -2,16 +2,11 @@
 
 import re
 
+from aiopioneer.const import AVRCodeStrMap
+
 from ..const import (
     SOURCE_TUNER,
     MEDIA_CONTROL_SOURCES,
-    SpeakerModes,
-    HDMIAudioModes,
-    HDMIOutModes,
-    PQLSModes,
-    AMPModes,
-    PanelLock,
-    DimmerModes,
     Zone,
 )
 from ..params import (
@@ -21,6 +16,59 @@ from ..params import (
     PARAM_SPEAKER_SYSTEM_MODES,
 )
 from .response import Response
+
+
+class SpeakerModes(AVRCodeStrMap):
+    """Speaker modes."""
+
+    code_map = {"0": "OFF", "1": "A", "2": "B", "3": "A+B"}
+
+
+class HDMIOutModes(AVRCodeStrMap):
+    """HDMI out modes."""
+
+    code_map = {"0": "ALL", "1": "HDMI 1", "2": "HDMI 2", "3": "HDMI (cyclic)"}
+
+
+class HDMIAudioModes(AVRCodeStrMap):
+    """HDMI audio modes."""
+
+    code_map = {"0": "AMP", "1": "PASSTHROUGH"}
+
+
+class PQLSModes(AVRCodeStrMap):
+    """PQLS modes."""
+
+    code_map = {"0": "OFF", "1": "AUTO"}
+
+
+class DimmerModes(AVRCodeStrMap):
+    """Dimmer modes."""
+
+    code_map = {
+        "0": "Brightest",
+        "1": "Bright",
+        "2": "Dark",
+        "3": "Off",
+        "9": "(cycle)",
+    }
+
+
+class AMPModes(AVRCodeStrMap):
+    """AMP modes."""
+
+    code_map = {
+        "0": "AMP ON",
+        "1": "AMP Front OFF",
+        "2": "AMP Front & Center OFF",
+        "3": "AMP OFF",
+    }
+
+
+class PanelLock(AVRCodeStrMap):
+    """Panel lock settings."""
+
+    code_map = {"0": "OFF", "1": "PANEL ONLY", "2": "PANEL + VOLUME"}
 
 
 class SystemParsers:
@@ -147,26 +195,6 @@ class SystemParsers:
         return parsed
 
     @staticmethod
-    def volume(
-        raw: str, _params: PioneerAVRParams, zone=Zone.Z1, command="VOL"
-    ) -> list[Response]:
-        """Response parser for zone volume setting."""
-        raw = "".join(filter(str.isnumeric, raw))  # select only numeric values from raw
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="volume",
-                property_name=None,
-                zone=zone,
-                value=int(raw),
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
     def mute(
         raw: str, _params: PioneerAVRParams, zone=Zone.Z1, command="MUT"
     ) -> list[Response]:
@@ -181,177 +209,6 @@ class SystemParsers:
                 property_name=None,
                 zone=zone,
                 value=(raw == "0"),
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def speaker_modes(
-        raw: str, _params: PioneerAVRParams, zone=None, command="SPK"
-    ) -> list[Response]:
-        """Response parser for speaker mode. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="speakers",
-                zone=zone,
-                value=SpeakerModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def hdmi_out(
-        raw: str, _params: PioneerAVRParams, zone=None, command="HO"
-    ) -> list[Response]:
-        """Response parser for HDMI out setting. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="hdmi_out",
-                zone=zone,
-                value=HDMIOutModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def hdmi_audio(
-        raw: str, _params: PioneerAVRParams, zone=None, command="HA"
-    ) -> list[Response]:
-        """Response parser for HDMI audio mode. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="hdmi_audio",
-                zone=zone,
-                value=HDMIAudioModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def pqls(
-        raw: str, _params: PioneerAVRParams, zone=None, command="PQ"
-    ) -> list[Response]:
-        """Response parser for PQLS mode. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="pqls",
-                zone=zone,
-                value=PQLSModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def dimmer(
-        raw: str, _params: PioneerAVRParams, zone=None, command="SAA"
-    ) -> list[Response]:
-        """Response parser for display dimmer. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="dimmer",
-                zone=zone,
-                value=DimmerModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def sleep(
-        raw: str, _params: PioneerAVRParams, zone=None, command="SAB"
-    ) -> list[Response]:
-        """Response parser for sleep timer. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="sleep",
-                zone=zone,
-                value=int(raw),
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def amp_status(
-        raw: str, _params: PioneerAVRParams, zone=None, command="SAC"
-    ) -> list[Response]:
-        """Response parser for AMP status. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="status",
-                zone=zone,
-                value=AMPModes[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def panel_lock(
-        raw: str, _params: PioneerAVRParams, zone=None, command="PKL"
-    ) -> list[Response]:
-        """Response parser for panel lock. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="panel_lock",
-                zone=zone,
-                value=PanelLock[raw],
-                queue_commands=None,
-            )
-        )
-        return parsed
-
-    @staticmethod
-    def remote_lock(
-        raw: str, _params: PioneerAVRParams, zone=None, command="RML"
-    ) -> list[Response]:
-        """Response parser for remote lock. (Zone 1 only)"""
-        parsed = []
-        parsed.append(
-            Response(
-                raw=raw,
-                response_command=command,
-                base_property="amp",
-                property_name="remote_lock",
-                zone=zone,
-                value=raw,
                 queue_commands=None,
             )
         )
